requirements.txt
pandas==1.5.0
pika==1.3.1
psycopg2==2.9.5
psycopg2_binary==2.9.3
python-dotenv==0.21.1
requests==2.28.1
ta==0.10.2
websocket_client==1.5.1
